+++ api/resources/vms/client/requests/index.ts
@@ -7,5 +7,5 @@ export type { PowerVmsRequest } from "./PowerVmsRequest.js";
 export type { ReinstallVmsRequest } from "./ReinstallVmsRequest.js";
 export type { ResetPasswordDto } from "./ResetPasswordDto.js";
 export type { ResizeVmDiskDto } from "./ResizeVmDiskDto.js";
-export type { ScaleVmsRequest } from "./ScaleVmsRequest.js";
+export type { ScaleVmDto } from "./ScaleVmDto.js";
 export type { UpdateHostnameVmsRequest } from "./UpdateHostnameVmsRequest.js";